lsm,selinux: pass flowi_common instead of flowi to the LSM hooks
commit 3df98d79215ace13d1e91ddfc5a67a0f5acbd83f upstream

As pointed out by Herbert in a recent related patch, the LSM hooks do
not have the necessary address family information to use the flowi
struct safely.  As none of the LSMs currently use any of the protocol
specific flowi information, replace the flowi pointers with pointers
to the address family independent flowi_common struct.

Reported-by: Herbert Xu <herbert@gondor.apana.org.au>
Acked-by: James Morris <jamorris@linux.microsoft.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>
Signed-off-by: Ehab Ababneh <ehab.ababneh@bytedance.com>