KVM: X86: Rename DR6_INIT to DR6_ACTIVE_LOW
commit 9a3ecd5e2aa10af18d0d5a055122d6cc0b0944c7 upstream.

DR6_INIT contains the 1-reserved bits as well as the bit that is cleared
to 0 when the condition (e.g. RTM) happens. The value can be used to
initialize dr6 and also be the XOR mask between the #DB exit
qualification (or payload) and DR6.

Concerning that DR6_INIT is used as initial value only once, rename it
to DR6_ACTIVE_LOW and apply it in other places, which would make the
incoming changes for bus lock debug exception more simple.

Intel-SIG: commit 9a3ecd5e2aa1 KVM: X86: Rename DR6_INIT to DR6_ACTIVE_LOW
Backport to add KVM support for Bus Lock Debug Exception

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>
Message-Id: <20210202090433.13441-2-chenyi.qiang@intel.com>
[Define DR6_FIXED_1 from DR6_ACTIVE_LOW and DR6_VOLATILE. - Paolo]
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
[ Aichun Shi: amend commit log ]
Signed-off-by: Aichun Shi <aichun.shi@intel.com>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>