x86/fpu/signal: Sanitize the xstate check on sigframe
commit 1258a8c896044564514c1b53795ba3033b1e9fd6 upstream.

Utilize the check for the extended state magic in the FX software reserved
bytes and set the parameters for restoring fx_only in the relevant members
of fw_sw_user.

This allows further cleanups on top because the data is consistent.

Intel-SIG: commit 1258a8c89604 x86/fpu/signal: Sanitize the xstate check on
sigframe.
Intel Advanced Matrix Extensions(AMX) is one of the key features on Sapphire
Rapids(SPR) and it depends on xsave, this patch is part of Thomas Gleixner's
x86 FPU code cleanup that upstream AMX implementation is building on, and is
also key to all code for being developed in upstream that depends on xsave.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Reviewed-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210623121457.277738268@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>