x86/fpu: Add size and mask information to fpstate
commit 248452ce21aeb08da2d2af23d88f890886bd379f upstream.

Add state size and feature mask information to the fpstate container. This
will be used for runtime checks with the upcoming support for dynamically
enabled features and dynamically sized buffers. That avoids conditionals
all over the place as the required information is accessible for both
default and extended buffers.

Intel-SIG: commit 248452ce21ae x86/fpu: Add size and mask information to
fpstate.
The second part of the effort to support AMX: Move register state into
a container struct.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20211013145322.921388806@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>