drivers/perf: hisi: Add TLP filter support
commit 17d573984d4d5ad73c7cb5edcf2024c585475b0c upstream.

The PMU support to filter the TLP when counting the bandwidth with below
options:

- only count the TLP headers
- only count the TLP payloads
- count both TLP headers and payloads

In the current driver it's default to count the TLP payloads only, which
will have an implicity side effects that on the traffic only have header
only TLPs, we'll get no data.

Make this user configuration through "len_mode" parameter and make it
default to count both TLP headers and payloads when user not specified.
Also update the documentation for it.

Reviewed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Yicong Yang <yangyicong@hisilicon.com>
Link: https://lore.kernel.org/r/20221117084136.53572-5-yangyicong@huawei.com
Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Slim6882 <yangjunshuo@huawei.com>
Signed-off-by: chenyi <chenyi211@huawei.com>
Signed-off-by: Jinhui Guo <guojinhui.liam@bytedance.com>