downstream: KVM:arm64:Add a kvm parameter to control guest wfi trapping
Currently, on GICv4+ system if a vCPU is only runnable process
on a pCPU and we are using direct injection of interrupts, the
guest WFI trapping will be turn off. So in host view, the vcpu
process's CPU util is almost 100%, which may confuse someone.
Let's add a kvm parameter to control guest wfi trapping.

Signed-off-by: Zhiqiang Ni <nizhiqiang1@huawei.com>
Signed-off-by: chenyi <chenyi211@h-partners.com>